Chore: SearchPreview을 SearchRecommendation으로 컴포넌트명 변경
- 보다 기능을 잘 설명할 수 있어 가독성을 높이는 컴포넌트명으로 변경
- 관련 import 문 수정
- 관련 Styled Component명 수정
